+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to Toisto will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Add support for the imperative grammatical mood. Closes [#618](https://github.com/fniessink/toisto/issues/618).
+
 ## 0.16.0 - 2024-03-30
 
 ### Fixed